refactor: map function to hooks
